start of work to refactor conda_oci_mirror to use oras
I am adding a global oras client, that way we can ensure checking
credentials once (and possibly defaulting to a dry run) and using
the client across different types of pushes (repo and packages).
I also added a "Pusher" class to oras since the actions of creating
layers and setting defaults were used between the two places. I
am currently maintaining the relatively flat structure of the
module, however this could be refactored in the future if needed
(I usually prefer to have a main module with the core functionality
but I will not push this preference on others! I need to next
refactor the cache_packages.py and then add some testing.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>