add support for getting package, info index.json for a package
the previous subdir_accessor.py had functions to get info, the index.json,
and the package (archive) when given a specific URI. These are now added
to the PackageRepo, meaning this class can now take a registry URI (host
and username) on init so it functions similarity to the previous class.
To be more consistent across the cli and classes, we now take the full
registry <host>/<username> as the registry argument instead of separating
them. It also did not make sense for the Mirror class to accept multiple
channels when we are generally just taking one. Finally, the pull/push
classes now are done with parallel tasks, and the structure of tasks
is more consistent. There are tests for these new functions, and
the upload of a latest for a package is removed. This means when
we want to download latest, we determine the latest based on
sorted versions and build numbers

Signed-off-by: vsoch <vsoch@users.noreply.github.com>